requirements.txt
langchain==0.2.7
langchain-community==0.2.7
langchain-core==0.2.12
langchain-groq==0.1.5
langchain-openai==0.1.7
langchain-pinecone==0.1.1
mysql-connector==2.2.9
nest-asyncio==1.6.0
ollama==0.2.1
pandas==2.2.0
pinecone-client==3.2.2
PyPDF2==3.0.1
sqlparse==0.5.0
streamlit==1.36.0
suds==1.1.2
word2vec==0.11.1
